+++ myokit/tests/test_simulation_opencl_vs_cvode.py
@@ -15,7 +15,6 @@ import numpy as np
 import myokit
 
 from shared import OpenCL_DOUBLE_PRECISION, DIR_DATA
-from shared import WarningCollector
 
 # Unit testing in Python 2 and 3
 try: