+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This page lists the main changes made to Myokit in each release.
   - [#687](https://github.com/MichaelClerx/myokit/pull/687) `Variable.set_rhs(None)` now removes an RHS, instead of setting it to `Number(0)`.
   - [#687](https://github.com/MichaelClerx/myokit/pull/687) Calling `SimulationOpenCL.set_conductance` will now override any conductances previously set with `set_connections`.
   - [#687](https://github.com/MichaelClerx/myokit/pull/687) `SimulationOpenCL.set_connections` no longer accepts `None` as a valid argument.
+  - [#687](https://github.com/MichaelClerx/myokit/pull/687) The `SimulationOpenCL` methods `set_conductance`, `conductance`, `set_connections`, `neighbours`, `set_paced_cells`, `set_paced_cell_list`, and `is_paced`, will now raise a `RuntimeError` if diffusion was disabled at construction.
   - [#689](https://github.com/MichaelClerx/myokit/pull/689) In Python 2, an `ImportError` is now raised if `myokit.ini` contains the sequence " ;" in any of its value (as this cannot be processed by Python 2's `ConfigParser`).
   - [#728](https://github.com/MichaelClerx/myokit/pull/728) The `LegacySimulation` class no longer suppresses CVODE warnings but passes them on to the `warnings` module (backported from new `Simulation` class).
 - Deprecated